Update models to be compatible with duckdb
Some of the models throw errors when run with duckdb due to a
circular CTE reference. This appears to be due to how the ref()
is rendered in the compiled SQL, so you end up with the order_items
CTE selecting from order_items.

Rather than allow circular CTEs, this change simply renames the CTE
in question and aliases selects out of it in order to minimize the
scope of these changes.

If we determine the recursive CTE approach is better, or if there's
a way to enable that for duckdb only, we can always update.