Switch profile to snowflake in dbt_project.yml
The CI job for dbt build depends on the snowflake profile, but
somewhere along the line we switched snowflake out for databricks,
which leads to persistent CI failures.

This puts snowflake back in place for now. We might consider making
this an env_var() property so we can run CI for different profile
configs.